kallithea-cli: use null handler to mute all console logging for ssh-serve
Augment logger_root to use the null handler when running ssh-serve. Set the log
level to CRITICAL to filter early and minimize the amount of logging sent to
the null handler.

It is however recommended to configure the system to use some logging facility
that can handle logging for SSH access from many different processes.