scm: select options should be tuples - not lists with 2 elements
We use tuples in all(?) other places, so we should do the same here.

Since lists can be unpacked as tuples, it currently works to use 2-lists. But
in a following change for "correct" handling of webhelpers.select when using
webhelpers2.select, we want to give special handling to tuples vs strings vs
Options. We don't want to add list to that list.