dev_requirements.txt: bump minimum pytest versions to a working set
Not all combinations of pytest-related packages seem to work.
Trial-and-error lead to these combinations of minimum versions.

Add a minimum bound to pytest-benchmark and pytest-localserver to help pip
in figuring out a suitable combination with pytest, and also to allow
testing minimum combinations in the future.

Drop pytest-runner, which should not be needed unless when running tests
directly from setup.py.