scripts/validate-commits: use direct /bin/bash interpreter
Use of '#!/bin/bash' vs '#!/usr/bin/env bash' can be argued multiple ways.
For example, see:
https://stackoverflow.com/questions/21612980/why-is-usr-bin-env-bash-superior-to-bin-bash/21613039

It seems to be more common to use the first approach. And we don't have a
specific reason to prefer the latter. Moreover, other scripts in this repo,
like 'make-release' and 'whitespacecleanup.sh' use the former too.