page: drop most paginate customizations - the bare implementation seems good enough
The main difference from this seems to be that redundant controls (such as
"previous" on first page) no longer will be shown.

default_link_tag is based on the base class implementation with minimal
changes, wrapping links in <li> as before, and with different handling of
current_page to emit markup as before.