scripts/make-release: install ldap and pam to fix isort instabilities
isort (triggered by scripts/whitespacecleanup.sh) needs to know which
modules are local and which are not, to separate them with a newline. If a
module cannot be found, it will be treated as local, apparently.

When ldap is not installed in the current environment, a difference was
created by isort in kallithea/bin/ldap_sync.py.